Add Application class to sample app so we can do lifecycle callbacks properly Break sample calls into private funs add debug and local debug flavors
