+++ sample/src/main/AndroidManifest.xml
@@ -1,11 +1,14 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
+        android:name="SampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Klaviyoandroidsdk">
+        android:theme="@style/Theme.Klaviyoandroidsdk"
+        android:usesCleartextTraffic="true"
+    >
         <activity
             android:name=".MainActivity"
             android:exported="true"
